lots of WIP, but stable enough to run - new oclModel that Fluids/Boids/Clouds inherits, they were pure opencl anyway - ocl static lib is removed, there is now only the physics lib which simplifies the dependencies internal to physics - using variadic templates to set the kernel inputs at class declaration level - target is: generic interface to UI through json provided by Model class, no implementation shared outside of Physics
