+++ physics/CMakeLists.txt
@@ -9,7 +9,8 @@ include_directories("utils")
 add_subdirectory("ocl")
 add_subdirectory("utils")
 
-target_link_libraries(physics PRIVATE ocl utils)
+# Not the physics/utils folder but the generic static utils lib
+target_link_libraries(physics PRIVATE utils)
 
 target_include_directories(physics INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
